[#1296] Add support for auto-provisioning.
Auto-Provisioning is documented in concept page *Device Provisioning*.
In the tenant configuration a new flag is added to enable or disable
auto-provisioning for each configured trusted CA.

If auto-provisioning is enabled for the trust anchor of the client
certificate used by a device for authenticating, the serialized device
certificate is put into the payload of the GET request to the Credentials
API to enable auto-provisioning in the Device Registry.

The GET request to the Credentials API can now return the status code 201.
The client implementation handles it in the same way as status code 200.

Signed-off-by: Abel Buechner-Mihaljevic <Abel.Buechner@bosch-si.com>