Add Hotrod based DeviceConnectionClient
The client implementation connects directly to an Infinispan data grid
to access device connection information. This client is an optimized
version of the existing client, skipping the Device Connection service
altogether. Note that this also means that the client can access all
data for all tenants and should therefore not be used with custom
protocol adapters.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>