[#1658] Fix Infinispan server clustering
The members of the Infinispan server cluster were not able to find each
other because of the wrong label being used to look up pods.

The Infinispan pods are now deployed as part of a StatefulSet which
allows them to use proper persistent storage if desired. Consequently,
the service for the data grid is now a headless service and can thus no
longer be accessed from outside of the Kubernetes cluster but only via
the Device Connection service.

Fixes #1658

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>