[#1661, #504] Include MQTT client ID in Credentials.get request
The MQTT adapter now always includes the device's MQTT client identifier
in the *client context* which is sent to the Credentials service in the
get request.

In order for this to work, the example Credentials service
implementation had to be changed so that it no longer asserts that a
device's credentials on record also contain the properties submitted in
the client context. This check doesn't seem to make much sense in the
context of the use case at hand anyway. The overall idea of including
the client context was not to have additional matching criteria for
credentials but to allow a Credentials service implementation to e.g.
use information from the client context as the device ID or
authentication identifier (as in the Kapua case)

Fixes #1661

Signed-off-by: Kai Hudalla <kai.hudalla@bosch-si.com>