Use HTTP based readiness/liveness probe
The Dispatch Router's standard AMQP port had been used for
readiness/liveness probing which resulted in the log files being
polluted with messages indicating that the peer (Kubernetes) did not
send proper AMQP frames.

In order to fix this, the Dispatch Router now listens on the standard
health check port for HTTP requests.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>