[#1761] Add AMQP adapter client. (#1765)
* [#1761] Add AMQP adapter client.

Device (or gateway) clients that connect to the AMQP adapter can be created
using AmqpAdapterClientFactory. The implementation reopens closed connections
and restores the receiver link for consuming commands afterwards. Sender links
are cached and removed from the cache when closed so that they are recreated
the next time they are retrieved from the factory.

Fixes #1761

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>