Manage life cycle of device registry beans from endpoints
The beans implementing the example device registry have been deployed as
Verticles in the past. This has the advantage of having their life cycle
being managed by the vert.x container (i.e. invoking their start/stop
methods).

However, it also means that we have no control over the event loop
context (thread) they will run on. For the example registry this is not
a big issue because the beans will be accessed from multiple threads
anyway (from the AMQP and the HTTP server verticles) and we already took
care of allowing for this by using ConcurrentMaps.

For registry implementations that keep the data in a back end store like
a DB, this approach is not desirable, because it would require us to
give up the "single thread only" programming model of vert.x.

Instead, the beans implementing the data access layer in the registry
now implement the Lifecycle interface and have their start/stop methods
being invoked as part of the AMQP/HTTP server verticle's life cycle.

For registries using an external persistent store, this requires
implementing the different APIs in separate beans which can then be
injected into just the AMQP server endpoints or the HTTP server
endpoints, thus making sure that they are always only accessed from a
single thread, even if the server verticles get deployed multiple times.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>