Use distinct service bean instance per endpoint
The service beans need to scale out with the number of AMQP and HTTP
server instances. In order to make sure that they run on the same event
loop thread as the server that invokes their operations, they must not
be deployed as Verticles but instead be controlled as part of the
endpoints life cycle.

The registry no longer uses the Spring Boot component scan mechanism but
instead creates the service beans explicitly in order to make the
process more transparent.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>