Fix a NPE in the TenantObject
According to the spec, the field "auto-provisioning-enabled" is not
mandatory. However the method TenantObject#addTrustAnchor throws a
NPE, while allowing a Boolean (vs boolean) parameter to be passed.

Additionally the method TenantObject#isAutoProvinsingEnabled() fails
with a NPE if the field is missing on the trust anchor.

This change also logs the full stack trace.