Remove obsolete get method
The Device Registration API does not define a get operation anymore.
Thus, the corresponding get method has been removed from tge
DeviceRegistration interface in the client module.

The Command & Control functionality in the Kerlink Lora provider has
been removed because it
a) erroneously relied on this method and
b) did not work properly anyway as it did not comply with Hono' C&C
addressing scheme.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>