[#1983] Run single Mongo DB instance for unit tests
The unit tests are now run using a single Mongo DB instance that is
started during the pre-integration-test phase and stopped in the
post-integration-test phase. The building of the Docker image has been
bould to the post-integration-test phase as well in order to not build
the image if any of the tests fail.

The downloaded Mongo DB artifacts are downloaded once only, cached
locally and re-used on subsequent builds.

Fixes #1983

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>