[#1023] Adding metric for connection attempts
In order to be able to fine-tune connection limits it is useful to know how many connections were actually rejected. This is now enabled by introducing the metric "hono.connections.attempts" and a corresponding "outcome" tag which holds the result of the connection attempt and hinting for a reason if applicable.

Signed-off-by: Florian Kaltner <florian.kaltner@bosch.io>