[#1023] Adding metric for connections attempts
In order to be able to fine-tune connection limits it is useful to know how many connections were actually rejected and why. For this purpose the hono.connections.attempts metric has been added. The metric supports specifying an outcome via a tag. For now the only supported outcome is adapter-connection-limit-exceeded.

Signed-off-by: Florian Kaltner <florian.kaltner@bosch.io>