[#1824] Improving message limit accuracy
Hono's default implementation for resource limits, which is based on Prometheus,caches the queried metrics until they hit a configurable expiry time. The limit is only exceeded when the Prometheus metrics reports so, which may take until the cache finally expires.

This is now improved by taking "local" values into account, i.e. the cache is updated without touching expiry of its entries. When it expires it will be synchronized with Prometheus again.

This fixes #1824

Signed-off-by: Florian Kaltner <florian.kaltner@bosch.io>