Add integration test for AMQP message exceeding max-message-size
The AMQP adapter is now being tested to reject messages exceeding the
adapter's configured max-message-size. The test can be enabled with a
feature flag as the underlying vertx-proton library does not (yet)
support this. Thus, the test is disabled for the time being.

See https://github.com/vert-x3/vertx-proton/issues/57

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>