[#2267] Add AMQP 1.0 based DeviceRegistrationClient implementation
Added implementations of the adapter client's DeviceRegistrationClient
interface which simply wraps the existing vertx-proton based
"legacy" RegistrationClientImpl.

In order to be able to use the legacy CommandTargetMapper, its direct
dependency on RegistrationClientFactory and
BasicDeviceConnectionClientFactory has been replaced with a
CommandTargetMapperContext interface which provides access to the
required functionality. Existing code using the CommandTargetMapper with
the legacy client will need to be adapted. For convenience, the
CommandTargetMapper.createContext() method can be used to get an
implementation based on RegistrationClientFactory and
BasicDeviceConnectionClientFactory.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>