[#2288] Fixing device level QoS forwarding of events for HTTP adapter.
The HTTP adapter suffered from a bug when a client sent an event and set a device level QoS to something different than 1 or set no device level QoS at all. In any case the event was sent with QoS 1 but the message property indicating the device level QoS to the northbound application (which should always be 1 as this is the only supported QoS for events) was wrong.

Signed-off-by: Florian Kaltner <florian.kaltner@bosch.io>