[#2029] Use Command Router in adapters and integration tests.
This adds a org.eclipse.hono.adapter.client.command.CommandConsumerFactory
implementation that uses the new Command Router component.
For the integration tests, a new maven profile 'command-router' is added to
let the tests run using the Command Router component.
The GitHub action workflow has been adapted to use that profile in the
test-run that uses the jdbc device registry (so that the other test-runs
still use the old command routing mechanism).

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>