[#2333] Add integration tests
Event messages are supposed to be queued by the Artemis broker when no north bound application is connected.
Once the north bound application connects those queued messages should be delivered. This was not working as
expected due to a bug described in the issue #2333. Integration tests are added to test the expected behavior.

Signed-off-by: Kartheeswaran Kalidass <kartheeswaran.kalidass@bosch.io>