Fix getContentType() returning empty string.
A not-set 'Content-Type' header caused an empty String
to be returned by getContentType() although 'null' was
documented in the javadoc.
Since the AMQP message anyway gets the
"application/octet-stream" content type in that case,
the getContentType() method now returns
"application/octet-stream" as the default (returning
'null' would have produced an NPE down the line).

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>