[#2363] Support inclusion of tenant ID in command topic
The MQTT adapter now allows clients to indicate whether they want the
target device's tenant and/or device IDs to be included in the topic
used when publishing commands.

This can be helpful for implementing (multi-tenant) protocol gateways
which can then determine a device's tenant and/or device ID from the
topic name.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>