Skip running unit tests on environments without Docker daemon
The JDBC and MongoDB unit tests which depend on Testcontainers for
running a DB container image, can now be skipped on environments by
setting the "noDocker" Maven property or activating the "no_docker"
Maven profile explicitly. This is necessary to allow the build process
to succeed on e.g. the Eclipse Foundation's JIRO build infrastructure
which doesn't provide access to Docker from within a build process.

The PostgreSQL image to use for running the tests can now be specified
using the Maven property "postgresql-image.name".

The license attributions for H2 and PostgreSQL have been added to the
legal module.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>