[#2403] Fix query for get credentials operation
The Mongo DB based Credentials service implementation failed to return
the credentials matching the given type and auth-id if multiple
credentials of the same type but with different auth-id values were
registered for a device.

The query has been changed to use the $elemMatch operator in conjunction
with the $ projection operator for specifying the (multiple) predicates
on the credentials type and auth-id.

See https://docs.mongodb.com/v4.2/reference/operator/projection/positional/#array-field-limitations
for an explanation.

Fixes #2403

Conflicts:
	site/homepage/content/release-notes.md
	tests/src/test/java/org/eclipse/hono/tests/registry/CredentialsApiTests.java

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>