[#2409] Remove creation of indices from start up process
The creation of indices had been done as part of the start up process of
the Mongo DB based registry service implementations. This causes the
start up to fail if the DB connection is not established (yet) when the
components are started. It is quite likely for this to happen in
scenarios where multiple components are started in parallel and compete
for the same resources.

The service implementations have been changed to perform the creation of
the indices independently from the startup process. The components now
also implement HealthCheckProvider and register a check which verifies
if the indices have already been created and, if not, trigger the
creation of the indices again.

This way, the component can start up successfully but remain in a
non-ready state until the indices have been created successfully.

Fixes #2409

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>