Refactor Command classes.
The new Command class now requires tenant/device information
to be set and returns this kind of information even if the
command is otherwise invalid. This makes it easier to choose
the command handler for a potentially invalid command.
Usage of the legacy command class inside the ProtonBasedCommand
has been removed.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>