Drop dependency on ProtocolAdpaterProperties
Mayn classes of the adapter-amqp module required a
ProtocolAdapterProperties instance in their constructors. However, some
classes will also be used outside the context of a protocol adapter. In
such cases, it seems odd to provide a properties instance which does not
actually represent the configuration of a protocol adapter.

The base classes for implementing AMQP based clients have been changed
to require simple booleans instead of the ProtocolAdapterProperties in
their constructors. The concrete client implementations have been
updated accordingly.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>