Register separate readiness check for each component instance
The Mongo DB based service implementation components register a
readiness check which triggers creation of the collection specific
indices. The successful creation of the indices is the condition that
determines the outcome of the readiness check.

In order to make sure that each component's readiness check is actually
considered in cases where multiple instances of the components are being
used, the readiness check's name now includes a UUID so that they do not
overwrite each other when registering the check with the
HealthCheckServer.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>