Decouple startup of AutoProvisioner from downstream connection
The AutoProvisioner's start up process was dependent on the
EventSender's establishing of its connection to the downstream messaging
network. This could lead to the registry container failing to start in
cases where the connection cannot be established quickly enough.

The AutoProvisioner's start up process has been changed to only trigger
the EventSender to connect to the messaging network and rely on the
EventSender's readiness check to indicate the establishment of the
connection.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>