[#8] Commit offsets when the Kafka consumer is interrupted.
If an error occurs while processing a message or when the consumer is stopped,
the offsets for the already processed messages from the current batch are now
committed. This is to prevent all messages from being processed again.
Especially in case of errors during message processing, the same messages could
otherwise be consumed again and again.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>