[#8] Add "retry" to AbstractAtLeastOnceKafkaConsumer.
The message handler passed into AbstractAtLeastOnceKafkaConsumer may throw a
ServerErrorException to indicate a transient error. In contrast to other
exceptions thrown by the message handler, the consumer will not be closed in
this case. Instead, the current offsets will be committed and the failed message
will be fetched again with the next poll operation.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>