[#2402] Use the messaging system configured on the tenant.
Protocol adapters start AMQP-based event senders *and* Kafka-based event
senders (if configuration  for both is provided). At the tenant a new
property is added to select which messaging to use. If this configuration
is present, that messaging is used to send events downstream.
The tenant configuration is currently only used for events. Telemetry and
Command and Control needs to be added as well.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>