Fix pushing of images
The Maven Docker plugin had been configured to always build and push an
image, even if the module was of type "pom" and did not produce any
code.

The plugin's execution is now tied to the image build process and
pushing of images is controlled by the "docker.push.skip" Maven property
which is true by default and set to false when the "docker-push-image"
profile is active.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>