[#2521] Let disconnect() abort an ongoing connection attempt.
The connect() method has been refactored so that a connection
attempt along with potential retries is handled in a separate
inner class named "ConnectionAttempt". The old implementation
allowed for separate connect() invocations to potentially lead
to separate, concurrent retry loops (if invoked while the other
attempt was in the reconnectDelay timer). This is not possible
any more. With the encapsulation in the ConnectionAttempt class,
an attempt can now be cancelled, aborting a potentially ongoing
reconnectDelay timer. This is used in the disconnect() and
shutdown() methods.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>