[#2551] Provide wrapper around transport specific client implementations
The MessagingClientSet class is taylored to be used by protocol adapters
which always require a TelemetrySender, EventSender and
CommandResponseSender for the configured transport. The logic to
determine the particular client implementation suitable for a given
tenant is contained in MessagingClients.

In order to allow plain services like device registry implementations to
take advantage of this logic for a single type of sender, this logic has
been refactored into the typed MessagingClient<? extends Lifecycle>
class in the client-common module.

This way, services like device registry implementations do not need to
depend on hono-adapter-base anymore but can still take advantage of
getting a client implementation appropriate for the tenant's configured
messaging system.

Fixes #2551

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>