[#2490] Support Kafka messaging in Quarkus based Adapters.
This adds support for messaging with Kafka to the Quarkus-based protocol
adapters. It uses the Quarkus Kafka Extension which provides two features:
a) it parses the Kafka configuration and b) it provides fixes to the original
Kafka client that are necessary to run it in GraalVM. The class
`KafkaRuntimeConfigProducer` overwrites the config parsing so that it
supports the Hono-specific prefixes for the Kafka configuration.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>