Run command handler in correct vert.x context.
Command processing in the adapters needs to be done
in the correct vert.x context. Otherwise there will
for example be concurrency issues in the AMQP adapter
concerning the AMQP connection to the device used for
sending the command. Since there is only *one*
KafkaBasedInternalCommandConsumer instance where
commands are received, command handlers explicitly
need to be run on the correct context. (That isn't
a problem with the ProtonBasedInternalCommandConsumer
since its (multiple) instances use the same context
as the corresponding adapter verticles.)

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>