Do not use Spring profile for enabling embedded cache
The Command Router component now uses the
HONO_COMMANDROUTER_CACHE_REMOTE_SERVERLIST configuration variable during
start up to determine if a remote or an embedded cache should be used.
This is in line with the Quarkus based variant of the component which
uses this approach as well instead of Quarkus runtime profiles which are
not quite the same as Spring profiles. In particular, it is not possible
to activate multiple profiles as in Spring but only a single profile can
be set.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>