[#2576] Support re-enabling of command routing after restart
The Command Router API has been extended with an operation to be used by
protocol adadpters for submitting tenant IDs for which the Command
Router should (re-)enable routing of commands after an unexpected
re-start. The command router will then try to (re-)create a command
consumer for the submitted tenant IDs.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>