[#2587] Support an in-memory database with the JDBC registry.
This adds support for a H2 in-memory database to the JDBC based
device registry. It is enabled by activating the application profile
"in-memory". With this profile a static configuration is used and
any provided properties with the prefixes "hono.registry.jdbc" or
"hono.tenant.jdbc" are ignored. Flyway executes the scripts in the
resources directory "db.migration" in their alphabetical order.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>