Refactor CoAP resources into separate classes
The resources used for uploading telemetry, event and command response
messages have been moved into separate classes. This reduces complexity
of the protocol adapter class and allows for easy testing of the
resources individually.

For this purpose, the ProtocolAdapter interface has been introduced as a
collection of methods providing collaborators and functionality that the
resources require for processing requests.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>