Add basic LwM2M registration support to CoAP adapter
The CoAP adapter has been extended with a "Resource Directory" resource
as used by LwM2M 1.0.2. It currently supports registration, updating of
a registration and de-registration of devices using binding modes U and
UQ only.

Upon registration the adapter creates a command consumer for the device
and sends an empty notification downstream which contains a TTD
reflecting the binding mode, i.e. the TTD will be the registered
lifetime for binding mode U and a fixed 20 seconds for mode UQ.

However, the adapter does not support forwarding commands to devices
yet.

The adapter also creates observations of LwM2M objects Device and
Firmware upon registration. In a future version the particular resources
to observe should be configurable at the device level.

Notifications received for the observations are being forwarded
downstream as telemetry or event messages, depending on the type of
object being observed.

The CoAP adapter's configuration properties have been extended with an
undocumented feature flag ("lwm2mEnabled") for enabling support for
LwM2M. The feature is disabled by default.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>