Fix for sending commands and receiving responses over Kafka
Currently the correlation id is stored in a collection after successfully sending a
command. There is a chance that the command response is received by the Kafka
consumer before the correlation id is even stored. This has been fixed now by storing
the correlation id before calling the send command method. If sending a command
fails, then the corresponding correlation id is removed from the collection.

Signed-off-by: Kartheeswaran Kalidass <kartheeswaran.kalidass@bosch.io>