[#2620] Add cipher suite tag to hono.connections.attempts metric
The metric used for reporting devices' connection attempts has been
extended with tag "cipher-suite" which takes the name of the cipher
suite used by a device's connection to a protocol adapter.

The AMQP and MQTT adapters have been updated to report the cipher suite
as part of connection attempt processing.

Fixes #2620

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>