Improve encapsulation of LimitedResource
The LimitedResource class has been extended with a method for
calculating the period of time that has passed since the start of the
current accounting period. This information is neede for example for
checking limits that have been defined for a tenant.

So far, the functionality for calculating this time period had been
implemented in PrometheusBasedResourceLimitChecks but it seems more
reasonable to implement it in LimitedResource which contains all the
required information in its properties already.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>