[#2654] Support persisting data to local file system
The Infinispan class implementing the file based data store has been
added to the native image build's reflection configuration so that it
will be included in the resulting native image.

The integration test's cache configuration has been changed to use file
system persistence in order to verify the embedded cache configuration
when running the tests with native images.

Fixes #2654

Conflicts:
	site/homepage/content/release-notes.md

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>