[#2084] Reduce load on Prometheus server when checking limits
The Prometheus based resource limit checks were prone to generating a
lot of load on a Prometheus server under certain conditions.
In particular, the load would spike, if a protocol adapter to which many
devices of many different tenants were connected would crash (or be shut
down) and the devices would immediately re-connect to another instance
of the protocol adapter. In this case, for the first device per tenant
up to three queries would be issued to Prometheus' query API, depending
on whether the tenant's current limited resource values were already
available in the cache or not. For multiple devices from different
tenants connecting in parallel, this could lead to a high load of
queries on the Prometheus server.

The limit checks have been changed to reduce (if not prevent) these
spikes:
1. The Java Executor used for running the queries has been changed from
the default (multi-threaded) implementation to a single-threaded
implementation. This results in one query being issued to Prometheus at
a time (per adapter instance).
2. When invoking any of the methods to check if a limit is reached, the
methods now immediately return FALSE (passing the check) if the required
limited resource value for the tenant is not yet in the cache.
Regardless of this behavior, a task to load the value into the cache is
scheduled using the Executor in the background.
3. The cache implementation now refreshes values in the background that
have reached half of their expiration time but are still of interest. The cache
returns the existing (possibly stale) value while the refresh task is
running, thus making sure that checks can always be performed based on
data from the cache.

In addition to these changes, the query to determine the used data
volume for the current accounting period has been slightly optimized to
reduce the amount of data that needs to be processed on the Prometheus
server.

Also increased the default cache timeout to from 15 to 60 seconds.

Fixes #2084

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>