Move tenant conversion test to an appropriate class
A unit test to verify conversion of a Tenant instance to TenantObject
is in FileBasedTenantServiceTest. This conversion utility method is scoped
not only to the file based registry. Moreover there is a plan to remove the
file-based device registry in the near future. This test has been now moved
to an appropriate test class.

Signed-off-by: Kartheeswaran Kalidass <kartheeswaran.kalidass@bosch.io>