[#2675] Adapt connection limit strategy to VM being used
The protocol adapters use the MemoryBasedConnectionLimitStrategy to
determine the number of connections that can be established with devices
based on the memory available to the adapter process.

The adapters currently assume to be running on a standard Java VM and
therefore assess the minimum amount of memory required for running to be
100MB (excluding memory allocated for device connections).

When using the Quarkus based native images, the memory footprint of the
adapters is significantly lower. This is now reflected in the MQTT and
AMQP adapters which first check the type of VM they are running on and
then adapt the minimum required memory accordingly.

Fixes #2675

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>