[#2735] Fix updating empty set of credentials
The registry implementations failed to reject a request to update a
device's empty set of credentials, e.g. right after the device has been
created, if the request contained a secret having an ID. In fact, the
registry implementations would accept such a request even if the secrets
did not contain a password/key etc at all but only the identifier.

The registries have been changed to now explicitly check if a device's
credentials on record are empty before merging in updated credentials.

Fixes #2735

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>