[#2736][#2592] Support receiving commands via both AMQP and Kafka.
If both AMQP and Kafka messaging systems are configured, then
the command router will receive command messages via both
systems.
The used messaging system is encoded in the command message
forwarded to the device (as part of the request id or the command
address in case of an AMQP device).
A command response message will then be forwarded using the
messaging system that the command was received in. If that
system isn't configured any more, the (tenant/global) default
messaging system will be used.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>