Improve Command Router C&C tracing.
This adds a FOLLOWS_FROM span reference from the
command handling trace started in the Command
Router to the trace where the command consumer
was created in the protocol adapter.
This is only enabled for short-lived command
consumers, i.e. those that get created in the
HTTP/CoAP adapters for ttd requests.

Signed-off-by: Carsten Lohmann <carsten.lohmann@bosch.io>