[#2110] Pass MongoClient into DAO constructors
The DAO implementations have been changed to no longer create the
MongoClient required for accessing the database themselves but to
instead have the client being passed in to the DAOs' constructors.

This makes it much easier to unit test the DAOs because they can be run
with a MongoClient mock.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>