[#2727][#2683] Make Email address configurable.
To avoid putting the email address to be used by Let's Encrypt into the
files, this has been made configurable. As the scripts already require
Helm to be installed, Helm's templating is used to create the yaml files
dynamically based on the Let's Encrypt API to be used and the email address.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>