Decouple mapped payload from original MQTT PUBLISH packet
The HTTP based payload mapper currently returns the original vert.x
Buffer containing the request payload if no mapping URL is configured
for a device. This makes the mapped payload prone to alterations to the
underlying Nettx ByteBuf that contains the MQTT PUBLISH packet.

The mapper has therefore been changed to create a copy of the original
Buffer instead, effectively de-coupling its content from the ByteBuf.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>