Fix caching key creation
The client context passed in from the device is being used as part of
the key in the Credentials cache. In particular, the hash code of the
JsonObject is being used. However, the hash code being generated depends
on how binary values are stored inside of the JsonObject. In particular,
it makes a difference if a binary value is stored as a raw byte array or
a string representing its Base64 encoding. In order to "normalize" the
JsonObject, the client context is now being marshalled and unmarshalled
again before creating its hash code.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>