[#2725] Add fix for TOC indentation in documentation.
This adds a patch from the upstream repository of the Hugo theme
that is already merged into the master branch of that repository
but no release has been made since then. So this commit adds the
patched CSS file. The file is intended to be removed from Hono's
repository once a newer release of the theme is used.

Signed-off-by: Abel Buechner-Mihaljevic <abel.buechner-mihaljevic@bosch.io>