[#2559] Make regex for validating hashed-password auth-ids configurable
The MongoDB and JDBC based registries have been changed to allow for the
configuration of a regular expression that is used by the
PasswordCredential class to validate the authentication identifier (user
name). The regular expression is read from a System property by the
PasswordCredential class because it is instantiated by the Jackson
framework when unmarshalling a JSON string and I can currently not think
of any other way of passing it in.

Fixes #2559

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>