[#2819] Allow multiple tenants to have empty CA array
The (partial) unique index defined on the tenant's trusted-ca array had
been defined using an $exists expression on the trusted-ca property.
This caused tenants with an empty trusted-ca array to also be indexed.

The partial index now uses an $exists expression on
trusted-ca.subject-dn instead which means that only documents will be
indexed that have a non-empty trusted-ca array.

Fixes #2819

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>