[#2819] Prevent tenants from having empty CA array
The (partial) unique index defined on the tenant's trusted-ca array is
defined using an $exists expression on the trusted-ca property.
This causes tenants with an empty trusted-ca array to also be indexed.

The Tenant class' trustedCertificateAuthorities field has been annotated
to only be included in its JSON representation when the collection is
not empty.

Fixes #2819

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>