Adapt environment variable names to Quarkus requirements
Quarkus uses SmallRye under the hood for binding environment variables
and/or Java system properties to configuration objects as described in
https://quarkus.io/guides/config-mappings. The VERBATIM naming strategy
that is used in Hono does not allow environment variables to contain an
underscore character for a change in case of the property name. For
example the property "useDefaultMetrics" of a configuration object
requires an environment variable "USEDEFAULTMETRICS" instead of
"USE_DEFAULT_METRICS" using this naming strategy.

The admin guides have been updated accordingly.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>