[#2878] Report connection attempts in Authentication Server
The Authentication Server is used by the Qpid Dispatch Router to perform
authentication of downstream applications. In order to detect potential
bottlenecks during connection establishment with such applications,
the Authentication Server has been extended to report the number of
attempts made by clients to establish an AMQP 1.0 connection to the
server.

Fixes #2878

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>