Support creating native images on rootless Docker
When running the container used for creating the native executable on a
local Docker daemon one of the project's target folder subdirectories is
bind-mounted into the container so that the native-image process can
read the jar file and write the native executable.

For this to work, the container process needs to have write privilege on
that folder. When running on a rootless Docker daemon, this is not the
case because the user set when the container is started has a different
user id than the user on the host system.

Using a remote docker based build solves this problem by means of
copying the jar file and the native executable hence and forth between
the host system and the Docker container instead of bind mounting the
directory.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>