[#2946] Use default vert.x cache directory
Hono's components have been using the "vertx.cacheDirBase" system
property to set the directory where vert.x should store files resolved
via the class path.

With vert.x 4 the semantics of this property's value has changed which
might lead to incompatibilities (see
https://github.com/eclipse-vertx/vert.x/issues/4168).

The system property is no longer set for Hono components because the
default behavior is to create a subfolder in the JVM's tmpdir to which
the running process will (usually) have write access.

Fixes #2946

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>