[#2702] Improve Kafka client class hierarchy
The AbstractKafkaBasedMessageSender's send* methods have been changed to
require an OpenTracing Span to be passed in which will not be finished
by the send method anymore. The subclasses and test cases have been
adapted accordingly. With this change the classes are better in line
with the approach used in Hono in gerneral, i.e. the component creating
a Span should also be responsible for finishing it.

Some (now) obsolete send* methods have been removed as well.

Signed-off-by: Kai Hudalla <kai.hudalla@bosch.io>